+++ common/src/main/java/com/binaris/wizardry/content/spell/earth/Fangs.java
@@ -11,7 +11,7 @@ import com.binaris.wizardry.api.content.util.BlockUtil;
 import com.binaris.wizardry.api.content.util.GeometryUtil;
 import com.binaris.wizardry.content.spell.DefaultProperties;
 import com.binaris.wizardry.core.AllyDesignation;
-import com.binaris.wizardry.core.integrations.accessories.EBAccessoriesIntegration;
+import com.binaris.wizardry.core.integrations.ArtifactChannel;
 import com.binaris.wizardry.setup.registries.EBItems;
 import com.binaris.wizardry.setup.registries.Elements;
 import com.binaris.wizardry.setup.registries.SpellTiers;
@@ -61,7 +61,7 @@ public class Fangs extends Spell {
     }
 
     protected boolean spawnFangs(CastContext ctx, Vec3 origin, Vec3 direction) {
-        boolean defensiveCircle = ctx.caster() instanceof Player caster && caster.isCrouching() && EBAccessoriesIntegration.isEquipped(caster, EBItems.RING_EVOKER.get());
+        boolean defensiveCircle = ctx.caster() instanceof Player caster && caster.isCrouching() && ArtifactChannel.isEquipped(caster, EBItems.RING_EVOKER.get());
         boolean flag = false;
 
         if (ctx.world().isClientSide) {