+++ common/src/main/java/com/binaris/wizardry/content/spell/necromancy/SummonZombie.java
@@ -6,7 +6,7 @@ import com.binaris.wizardry.api.content.spell.internal.SpellModifiers;
 import com.binaris.wizardry.api.content.spell.properties.SpellProperties;
 import com.binaris.wizardry.content.spell.DefaultProperties;
 import com.binaris.wizardry.content.spell.abstr.MinionSpell;
-import com.binaris.wizardry.core.integrations.accessories.EBAccessoriesIntegration;
+import com.binaris.wizardry.core.integrations.ArtifactChannel;
 import com.binaris.wizardry.setup.registries.EBItems;
 import com.binaris.wizardry.setup.registries.Elements;
 import com.binaris.wizardry.setup.registries.SpellTiers;
@@ -27,7 +27,7 @@ public class SummonZombie extends MinionSpell<Zombie> {
 
     @Override
     protected Zombie createMinion(Level world, @Nullable LivingEntity caster, SpellModifiers modifiers) {
-        if (caster instanceof Player player && EBAccessoriesIntegration.isEquipped(player, EBItems.CHARM_MINION_VARIANTS.get())) {
+        if (caster instanceof Player player && ArtifactChannel.isEquipped(player, EBItems.CHARM_MINION_VARIANTS.get())) {
             return new Husk(EntityType.HUSK, world);
         } else {
             return super.createMinion(world, caster, modifiers);