+++ common/src/main/java/com/binaris/wizardry/setup/registries/EBTags.java
@@ -12,6 +12,13 @@ public final class EBTags {
     public static final TagKey<Item> RING_ACCESSORIES = createItemTag("accessories", "ring");
     public static final TagKey<Item> CHARM_ACCESSORIES = createItemTag("accessories", "charm");
     public static final TagKey<Item> NECKLACE_ACCESSORIES = createItemTag("accessories", "necklace");
+    public static final TagKey<Item> RING_CURIOS = createItemTag("curios", "ring");
+    public static final TagKey<Item> CHARM_CURIOS = createItemTag("curios", "charm");
+    public static final TagKey<Item> NECKLACE_CURIOS = createItemTag("curios", "necklace");
+    public static final TagKey<Item> MAIN_HAND_RING_TRINKETS = createItemTag("trinkets", "hand/ring");
+    public static final TagKey<Item> GLOVE_TRINKETS = createItemTag("trinkets", "hand/glove");
+    public static final TagKey<Item> NECKLACE_TRINKETS = createItemTag("trinkets", "chest/necklace");
+
     public static final TagKey<Block> GILDED_WOOD_BLOCK = createBlockTag("gilded_wood");
 
     private EBTags() {